Some functions are added. Python Wrapper is added.

+++ src/Transport.cpp
@@ -32,8 +32,8 @@ int Transport::SendPacket(unsigned char opCode,
 int Transport::ReceiveData(unsigned char *buffer, unsigned int requestSize, unsigned int* readBytes)
 {
 	unsigned char* data = new unsigned char[requestSize];
-	while (m_pSerialPort->GetSizeInRxBuffer() <= requestSize) {
-
+	while ((unsigned int)m_pSerialPort->GetSizeInRxBuffer() <= requestSize) {
+		
 	}
 
 	*readBytes = m_pSerialPort->Read(buffer, requestSize);
